Some features added

+++ src/shoesshop/Customer.java
@@ -4,10 +4,20 @@ public class Customer extends User{
     protected double cash;
     protected void increaseCash(double sum){
         this.cash += sum;
+        System.out.println("Your balance: "+ this.cash);
+        System.out.println();
     }
-    protected void buy(double sum, ShoesManager manager,int id){
-        this.cash += sum;
-        manager.deleteItem(id);
+    protected void buy(ShoesManager manager,int id){
+        Double sum = manager.findSumById(id);
+        if(sum >0 ) {
+            this.cash += sum;
+            manager.deleteItem(id);
+        }
+        else {
+            System.out.println("No shoes with such ID");
+
+        }
+        System.out.println();
     }
 
 }
